+++ src/index.ts
@@ -1,4 +1,7 @@
 import { Element } from './classes/Element'
 import { Style } from './classes/Style'
 
-export default { Element, Style }
+export default {
+  Element,
+  Style
+}