+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## 0.0.2
+
+### Features
+
+- Added Element content and style options in constructor.
+
+### Changes
+
+- Implemented the `Instruction` class.
+- Improved `Element` logs.
+
 
 ## 0.0.1 (Dec 6, 2021)